+++ CHANGELOG.md
@@ -7,10 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+_No unreleased changes._
+
+## [1.9.0] - 2022-05-09
+
 ### Added
 
-- `--on-pull` option to `zmodule`, which allows setting a command that is always triggered
-  after the module is installed or updated.
+- `--on-pull` option to `zmodule`, which allows setting a command that is always
+  triggered after the module is installed or updated.
 
 ## [1.8.0] - 2022-01-25
 
@@ -274,7 +278,8 @@ Take your time to review the updated [README.md] and the changes listed below.
 [termtitle]: https://github.com/zimfw/termtitle
 [s1ck94]: https://github.com/zimfw/s1ck94
 
-[Unreleased]: https://github.com/zimfw/zimfw/compare/v1.8.0...HEAD
+[Unreleased]: https://github.com/zimfw/zimfw/compare/v1.9.0...HEAD
+[1.9.0]: https://github.com/zimfw/zimfw/compare/v1.8.0...v1.9.0
 [1.8.0]: https://github.com/zimfw/zimfw/compare/v1.7.0...v1.8.0
 [1.7.0]: https://github.com/zimfw/zimfw/compare/v1.6.2...v1.7.0
 [1.6.2]: https://github.com/zimfw/zimfw/compare/v1.6.1...v1.6.2